refactor: re-write RAMSTKPanel to use new RAMSTKBaseWidget
This also moves the panel module into a panels package with each panel type in it's own file.

+++ src/ramstk/views/gtk3/widgets/plot.py
@@ -199,8 +199,8 @@ class RAMSTKPlot(RAMSTKBaseWidget):
         label: str,
         properties: PlotProperties,
         set_x: bool = True,
-        x_pos: int = 0,
-        y_pos: int = 0,
+        x_pos: float = 0.0,
+        y_pos: float = 0.0,
     ) -> matplotlib.text.Text:
         """Make the abscissa or ordinate label.
